requirements.txt
# Framework principal
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.4.2
python-multipart>=0.0.6

# Seguridad
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# Base de datos (elige según tu preferencia)
sqlalchemy>=2.0.23
# pymongo>=4.6.0  # Descomenta si prefieres MongoDB

# Utilidades
httpx>=0.25.0  # Para hacer solicitudes HTTP
tenacity>=8.2.3  # Para reintentos
tiktoken>=0.5.1  # Para contar tokens (OpenAI)

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1

# Logging y monitoreo
loguru>=0.7.2

# CORS
# starlette-cors